Support reading back theme color from an ooxml document
ooxml supports theme colors and tint/shade value that additionally
changed the theme color. Read back which theme color + tint/shade
value was applied in the resulting color and add this attributes
as properties to be used by writer.
In sidebar theme panel the changing the theme colors now doesn't
takes this into account and changes the colors correctly.

Change-Id: I6703e86b1fc6b2ba07f3023ec48e619aec961ff1