xmlsec: handle missing EqualDistinguishedNames
This is just a hack. A more correct fix would be to condition
the whole xmlsecurity support in the various LO modules.

The non-NSS build now fails, because it misses the
EqualDistinguishedNames, which is non-trivial to implement.

This is the consequence of commit
1d3da3486d827dd5e7a3bf1c7a533f5aa9860e42 "xmlsecurity: add
EqualDistinguishedNames()" and commit
5af5ea893bcb8a8eb472ac11133da10e5a604e66 "xmlsecurity: improve
handling of multiple certificates per X509Data".

FWIW: one could also use OpenSSL i2d_X509_NAME and X509_NAME_cmp
to implement this functionality, but it looks like it requires
some ANS.1 input to begin with (maybe i2s_ASN1_* can be used).
And then you would still miss some certificate store integration.

Change-Id: I0600f76a5cb4379db428f67cb136b8567dfcc51e